fix windows compilation
git-svn-id: http://svn.cern.ch/guest/lhcb/Boole/trunk@27792 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: f48f80f3c6d8edc4439b51b4917f66ab1fb02ebc

+++ Boole/ST/STDigiAlgorithms/src/STCommonModeSim.cpp
@@ -114,7 +114,7 @@ void STCommonModeSim::processPort(STDigits::const_iterator& start,
 
   unsigned int nMissed = LHCbConstants::nStripsInPort - (end - start);
   double noiseCounts = m_sigNoiseTool->noiseInADC((*start)->channelID());
-  double ranNoise = (nMissed* m_cmNoise) + m_gaussDist->shoot()*noiseCounts *sqrt(nMissed);
+  double ranNoise = (nMissed* m_cmNoise) + m_gaussDist->shoot()*noiseCounts *sqrt((double)nMissed);
 
   // make tmp list of charges
   std::vector<double> startCharge;
